fix: free still reachable and organizing functions
